add basic skeleton (with hook)
