+++ src/shared/hooks/withSkeleton.tsx
@@ -1,0 +1,34 @@
+import { Skeleton } from "../ui/Skeleton";
+
+type SkeletonType = "settings" | "list";
+
+// interface HookProps<P> {
+//   Component: ;
+//   ; //№Более конкретно уточнить тип до литералов
+//  ;
+//   //   direction?: string; //№Тоже самое
+// }
+
+interface ChildProps {
+  isLoading: boolean;
+}
+
+const withSkelton = <P extends object>(
+  Component: React.ComponentType<P>,
+  type?: SkeletonType,
+  count?: number
+) => {
+  const WithSkeleton = (props: ChildProps & P) => {
+    const { isLoading, ...otherProps } = props;
+
+    if (isLoading) {
+      return <Skeleton type={type} count={count} />;
+    }
+
+    return <Component {...(otherProps as P)} />;
+  };
+
+  return WithSkeleton;
+};
+
+export default withSkelton;
